+++ server/src/main/java/com/server/app/dto/request/RegionUpdateRequest.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class RegionUpdateRequest {
-    private Short regionId;
+    private Long regionId;
     private String regionDescription;
 }
