+++ server/src/main/java/com/server/app/dto/response/ProductDto.java
@@ -12,8 +12,8 @@ import lombok.NoArgsConstructor;
 public class ProductDto {
     private Long productId;
     private String productName;
-    private Short supplierId;
-    private Short categoryId;
+    private Long supplierId;
+    private Long categoryId;
     private String quantityPerUnit;
     private Float unitPrice;
     private Integer discontinued;
